+++ engine/report/sc_report_html_player.cpp
@@ -1904,7 +1904,7 @@ void print_html_sample_sequence_string_entry(
   for ( size_t b = 0; b < data.buff_list.size(); ++b )
   {
     buff_t* buff = data.buff_list[ b ].first;
-    int stacks   = data.buff_list[ b ].second[0];
+    int stacks   = gsl::narrow_cast<int>(data.buff_list[ b ].second[0]);
     if ( !buff->constant )
     {
       os.printf( "\n%s", buff->name() );
@@ -1994,7 +1994,7 @@ void print_html_sample_sequence_table_entry(
   for ( size_t b = 0; b < data.buff_list.size(); ++b )
   {
     buff_t* buff = data.buff_list[ b ].first;
-    int stacks   = data.buff_list[ b ].second[0];
+    int stacks   = gsl::narrow_cast<int>(data.buff_list[ b ].second[0]);
 
     if ( !buff->constant )
     {
